+++ src/components/Answer.tsx
@@ -1,44 +1,16 @@
-import React, { ChangeEvent, useState } from 'react';
+import React, { useState } from 'react';
 
-function Answer() {
-
-    const [question, setQuestion] = useState('')
-
-    const inputHandler = (e: React.ChangeEvent<HTMLInputElement>) => {
-        const target = e.target as HTMLInputElement;
-        setQuestion(target.value)
-    }
-
-    const clickHandler = (e: React.MouseEvent<HTMLButtonElement>) => {
-        e.preventDefault();
-        console.log('button clicked')
-        setRendererQuestion(question)
-
-    }
-
-    const [renderedQuestion, setRendererQuestion] = useState('')
+function Answer(props: any) {
 
 
 
     return (
         <>
-            <h1>This is a answer box</h1>
-            <form>
 
-                <label>
-                    Please enter your question:
-                    <input type="text" onChange={inputHandler} />
-
-                </label>
-                <button className='tell_me_the_answer' onClick={clickHandler}>
-                    Find the future
-                </button>
-            </form>
-
-            {renderedQuestion && (
+            {props.renderedQuestion && (
                 <div>
                     <h2>Your Question was:</h2>
-                    <p>{renderedQuestion}</p>
+                    <p>{props.renderedQuestion}</p>
                 </div>
             )}
         </>
